+++ data/savedTasks.txt
@@ -4,6 +4,5 @@ E | 1 | project meeting | 06/08/2019 14:00 | 06/08/2019 16:00
 T | 0 | join sports club
 T | 1 | cry
 D | 0 | submit stuff | 30/08/2024 16:00
-T | 0 | wipe tears
 D | 0 | buy food | 01/09/2024 23:59
 E | 0 | test | 02/09/2024 09:00 | 02/09/2024 11:00
