+++ src/main/java/barcus/command/AddCommand.java
@@ -1,3 +1,5 @@
+package barcus.command;
+
 public abstract class AddCommand extends Command {
     protected String description;
