+++ src/main/java/barcus/command/AddDeadlineCommand.java
@@ -1,25 +1,28 @@
+package barcus.command;
+
+import barcus.tasklist.TaskList;
+import barcus.ui.Ui;
+import barcus.storage.Storage;
+import barcus.exception.BarcusException;
+import barcus.task.*;
 import java.time.format.DateTimeParseException;
 
-public class AddEventCommand extends AddCommand {
-    protected String from;
-    protected String to;
+public class AddDeadlineCommand extends AddCommand {
+    protected String by;
 
-    public AddEventCommand(String description, String from, String to) {
+    public AddDeadlineCommand(String description, String by) {
         super(description);
-        this.from = from;
-        this.to = to;
+        this.by = by;
     }
-
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws BarcusException {
         try {
-            Task t = new Event(this.description, this.from, this.to);
+            Task t = new Deadline(this.description, this.by);
             tasks.addTask(t);
             ui.talk("Added task: " + t + "\nThere are " + tasks.getLength() + " task(s) in the list.");
         } catch (DateTimeParseException e) {
             throw new BarcusException("please format date as dd/MM/yyyy HH:mm");
         }
-
     }
 
     @Override
