+++ src/main/java/barcus/command/AddTodoCommand.java
@@ -1,3 +1,10 @@
+package barcus.command;
+
+import barcus.tasklist.TaskList;
+import barcus.ui.Ui;
+import barcus.storage.Storage;
+import barcus.task.*;
+
 public class AddTodoCommand extends AddCommand {
     public AddTodoCommand(String description) {
         super(description);
