+++ src/main/java/barcus/command/ExitCommand.java
@@ -1,3 +1,10 @@
+package barcus.command;
+
+import barcus.tasklist.TaskList;
+import barcus.ui.Ui;
+import barcus.storage.Storage;
+import barcus.exception.BarcusException;
+
 public class ExitCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws BarcusException {
