+++ src/main/java/barcus/command/MarkCommand.java
@@ -1,18 +1,21 @@
-public class DeleteCommand extends Command {
+package barcus.command;
+
+import barcus.tasklist.TaskList;
+import barcus.ui.Ui;
+import barcus.storage.Storage;
+
+public class MarkCommand extends Command {
     private int pos;
 
-    public DeleteCommand(int pos) {
+    public MarkCommand(int pos) {
         this.pos = pos;
     }
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (pos > 0 && pos <= tasks.getLength()) {
 //                        tasks.get(pos - 1).markDone();
-//                        talk("Good job! Have marked as done: " + tasks.get(pos - 1));
-            Task temp = tasks.deleteTask(pos - 1);
-//                        curr--;
-            ui.talk("Removed task: " + temp + "\nThere are "
-                    + tasks.getLength() + " task(s) in the list.");
+            tasks.markTask(pos - 1);
+            ui.talk("Good job! Have marked as done: " + tasks.getTaskString(pos - 1));
         } else {
             ui.showError("please choose a number between 1 and " + tasks.getLength());
         }
