+++ src/main/java/barcus/command/UnknownCommand.java
@@ -1,3 +1,9 @@
+package barcus.command;
+
+import barcus.tasklist.TaskList;
+import barcus.ui.Ui;
+import barcus.storage.Storage;
+
 public class UnknownCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
