+++ src/main/java/barcus/exception/BarcusException.java
@@ -1,3 +1,5 @@
+package barcus.exception;
+
 public class BarcusException extends Exception {
     public BarcusException() {}
