+++ src/main/java/barcus/task/Event.java
@@ -1,3 +1,5 @@
+package barcus.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
