+++ src/main/java/barcus/task/Task.java
@@ -1,3 +1,5 @@
+package barcus.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
