+++ src/main/java/barcus/task/Todo.java
@@ -1,3 +1,5 @@
+package barcus.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
