+++ src/main/java/barcus/ui/Ui.java
@@ -1,3 +1,5 @@
+package barcus.ui;
+
 import java.util.Scanner;
 
 public class Ui {
